# This read thread makes the Neato driver much faster