I was able to the driver much more responsive with improved read thread on my Raspberry PI.

`ubuntu@ubuntu:~$ ros2 topic hz odom`
`WARNING: topic [odom] does not appear to be published yet`
`average rate: 10.074`
`	min: 0.002s max: 0.401s std dev: 0.09989s window: 12`
`average rate: 10.059`
`	min: 0.002s max: 0.401s std dev: 0.07217s window: 23`
`average rate: 10.072`
`	min: 0.002s max: 0.401s std dev: 0.05941s window: 34`
`average rate: 10.011`
`	min: 0.002s max: 0.401s std dev: 0.05229s window: 44`
`ubuntu@ubuntu:~$ ros2 topic hz scan`
`average rate: 10.046`
`	min: 0.093s max: 0.113s std dev: 0.00473s window: 12`
`average rate: 9.941`
`	min: 0.093s max: 0.113s std dev: 0.00542s window: 22`
`average rate: 9.977`
`	min: 0.092s max: 0.113s std dev: 0.00513s window: 33`
`average rate: 9.954`
`	min: 0.092s max: 0.113s std dev: 0.00477s window: 43`
`ubuntu@ubuntu:~$`